jmavsim: advance time when we receive control data
This changes the behavior to only advance the "faked" time when we have
received control data (HIL_ACTUATOR_CONTROL).

Currently we run the update call twice per new sensor values in order to
make sure we receive an update from px4 before we want to send the next
sensor message.
This is done using the `slowDownCounter`.

Also, this separates getMillis into getRealMillis and getSimMillis to
make a clear distinction.