Simulator: make lockstep an option
This adds a CLI argument to enable lockstep. This should restore HITL in
its original working shape.

The problem was mainly that we used the simulator time instead of the
computer time and also we ran update twice which is required for
lockstep. One step is to receive messages from the autopilot, and the
other step is to send new messages out.

This change will require that we use the cli flag on the PX4 side.